Improve handling of max-speed below minimum
When the client's requested speed was below the minimum speed based on the clock
configuration, the driver would select illegal values for the BAUD_RATE_DIV
field and set a reserved bit which the TRM says should stay 0.  The resulting
behavior was that received bytes would be shifted down one bit and have their
MSB set (so 0x0A would become 0x85).

This change stops at the lowest speed the controller can acheive, which may be
higher than the client's requested speed but is more likely to work.  The speed
calculation is also run if the BAUD_RATE_DIV field is 0, which is not supported
according to the TRM.